Remove Regridder module
Remove all unused functions from Regridder module.
Used functions relocated to Utilities and FieldExchanger
modules. Docs and tests appropriately updated.

Re-add ClimaUtilities as dependancy after rebase

ClimaUtilities needs to be re-added after rebasing,
because it is now used for the ouput paths

Make suggested changes
