+++ Project.toml
@@ -6,12 +6,9 @@ version = "0.1.1"
 [deps]
 ClimaComms = "3a4d1b5c-c61d-41fd-a00a-5873ba7a1b0d"
 ClimaCore = "d414da3d-4745-48bb-8d80-42e94e092884"
-ClimaCoreTempestRemap = "d934ef94-cdd4-4710-83d6-720549644b70"
 ClimaUtilities = "b3f4f4ca-9299-4f7f-bd9b-81e1242a7513"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
-JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
 Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
-NCDatasets = "85f8d34a-cbdd-5861-8df4-14fed0d494ab"
 SciMLBase = "0bca4576-84f4-4d90-8ffe-ffa030f20462"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 SurfaceFluxes = "49b00bb7-8bd4-4f2b-b78c-51cd0450215f"
@@ -20,12 +17,9 @@ Thermodynamics = "b60c26fb-14c3-4610-9d3e-2d17fe7ff00c"
 [compat]
 ClimaComms = "0.5.6, 0.6"
 ClimaCore = "0.14.19"
-ClimaCoreTempestRemap = "0.3"
 ClimaUtilities = "0.1.14"
 Dates = "1"
-JLD2 = "0.4, 0.5"
 Logging = "1"
-NCDatasets = "0.11, 0.12, 0.13, 0.14.2"
 SciMLBase = "1, 2"
 StaticArrays = "1.5"
 SurfaceFluxes = "0.11, 0.12"
