+++ experiments/ClimaEarth/run_cloudy_slabplanet.jl
@@ -24,7 +24,7 @@ import ClimaCore as CC
 # ## Coupler specific imports
 import ClimaCoupler
 import ClimaCoupler:
-    ConservationChecker, Checkpointer, FieldExchanger, FluxCalculator, Interfacer, Regridder, TimeManager, Utilities
+    ConservationChecker, Checkpointer, FieldExchanger, FluxCalculator, Interfacer, TimeManager, Utilities
 
 import ClimaUtilities.ClimaArtifacts: @clima_artifact
 import ClimaUtilities.SpaceVaryingInputs: SpaceVaryingInput
