+++ src/ClimaCoupler.jl
@@ -8,7 +8,6 @@ module ClimaCoupler
 include("Interfacer.jl")
 include("Utilities.jl")
 include("TimeManager.jl")
-include("Regridder.jl")
 include("ConservationChecker.jl")
 include("FluxCalculator.jl")
 include("FieldExchanger.jl")
